Even more improvements
