+++ elastic-synthetics/src/index.ts
@@ -2,3 +2,9 @@ require('source-map-support').install();
 
 export * from './dsl';
 export * from './runner';
+
+import { run } from './runner' ;
+
+if (require.main === module)  {
+    run()
+}
