+++ elastic-synthetics/src/runner.ts
@@ -1,11 +1,30 @@
-#!/usr/bin/env node
-
-require('source-map-support').install();
-import { runJourneys } from './runner'
+import { state, journey, step } from './dsl';
+import * as playwright from 'playwright';
 import { Command } from 'commander';
 import { readFileSync } from 'fs';
-import { journey, step, state } from './dsl';
 import { createInterface as createReadlineInterface } from 'readline'
+
+
+const browserType = 'chromium'
+
+export const runJourneys = async (suiteParams: any) => {
+    console.debug("Running with suite params", suiteParams);
+    const browser = await playwright[browserType].launch({ headless: false });
+    const context = await browser.newContext()
+    const page = await context.newPage()
+    console.log(`Found ${state.journeys.length} journeys`)
+    for (let i = 0; i < state.journeys.length; i++) {
+        const journey = state.currentJourney = state.journeys[i]
+        console.log(`Journey: ${journey.options.name}`)
+        for (let j = 0; j < journey.steps.length; j++) {
+            const step = journey.steps[j]
+            console.log(`Step: ${step.name}`)
+            await step.callback(page, suiteParams, {context, browser})
+        }
+    }
+
+    await browser.close()
+}
 
 export const run = async (aroundHook: (runSteps: (suiteParams: any) => void, environment: string, suiteParams: any) => void = (r) => { r({}) }) => {
     const program = new Command;
@@ -20,6 +39,7 @@ export const run = async (aroundHook: (runSteps: (suiteParams: any) => void, env
         .action(async (file, options) => {
             if (options.suiteParams) {
                 const parsedSuiteParams = JSON.parse(options.suiteParams);
+                console.log("SPARAMS", parsedSuiteParams)
                 argSuiteParams = {...argSuiteParams, ...parsedSuiteParams};
             }
 
@@ -61,5 +81,3 @@ const loadInlineScript = (source, argSuiteParams) => {
     })
     console.log("r STEPS")
 }
-
-run()
