+++ sample-app/journeys/sample-inline-journey.js
@@ -1,16 +1,16 @@
-step('Go to home page', async (params) => {
+step('Go to home page', async (page, params) => {
     await page.goto(params.homepage)
 })
 
-step('Go to login page', async (params) => {
+step('Go to login page', async (page, params) => {
     await page.click('a')
 })
 
-step('Enter username and password', async () => {
+step('Enter username and password', async (page) => {
     await page.fill('input[name=username]', 'hamid')
     await page.fill('input[name=password]', 'test-pass')
 })
 
-step('submit form', async () => {
+step('submit form', async (page) => {
     await (await page.$('form')).evaluate(form => form.submit())
 })
